validate.py
import pytz
import string
import random
import json
import logging
import binascii
import dateutil.parser
import hmac
import os
import requests

from requests.exceptions import ConnectionError
from hashlib import sha1
from datetime import datetime

def validate_request(request, SCALR_SIGNING_KEY):
    if 'X-Signature' not in request.headers or 'Date' not in request.headers:
        logging.debug('Missing signature headers')
        return False
    date = request.headers['Date']
    body = request.data
    expected_signature = binascii.hexlify(hmac.new(SCALR_SIGNING_KEY, body + date, sha1).digest())
    if expected_signature != request.headers['X-Signature']:
        logging.debug('Signature does not match')
        return False
    date = dateutil.parser.parse(date)
    now = datetime.now(pytz.utc)
    delta = abs((now - date).total_seconds())
    return delta < 300
